requirements.txt
fastapi~=0.116.2
uvicorn~=0.35.0
pydantic~=2.11.9
httpx~=0.28.1
python-dotenv~=1.1.1
langchain~=0.3.27
langgraph~=0.6.7
elasticsearch~=9.1.1
tenacity~=9.1.2
langchain-openai~=0.3.33